Adding comparison operators for series_vector iterators
One of the reasons I had to eliminate the casting operator from sv
iterators was because the compiler was very eager to use it to get
pointers that it could then use in pointer comparisons. This kind of
worked with sv_iterator but of course reverse_sv_iterator was just very
wrong. So now without the casting operators, the iterators are failing
tests for comparisons. I've fixed this now.