Preliminary implementation of frame::corr()
This is a really rough implementation of frame::corr(). For one thing,
it's probably pretty slow - the iterators will be pretty fast, but I
haven't looked at the assembly code generated and more importantly I
haven't SIMD-ized this at all, so it won't be super-fast on large
datasets. Maybe, given that the iterators are just pointers and the
compiler knows all the types maybe it can do something clever, but I
need to profile this and look at what it's producing at the instruction
level.

Also, it has to be said that this won't be MP-aware either - a
correlation will only be done on a single core and that's how things
will be until I think about this a bit more. In a way, mainframe, with
it's copy-on-write semantics for columns, is kind of ideal for MP,
as long as you're very careful about using const frame<>'s and you join
on worker tasks/threads before mutating the original, but this is kind
of adhoc and provides the user with basically no protection at all. It
would be super-easy to wreck things in a worker thread/task because
mainframe has no exclusion locks for multiprocessing.

What should this look like? Exclusive access to a series_vector? A
series? Part of a series_vector? I don't think it's that useful to be
able to distribute a single correlation on a couple of columns to
multiple cores, but if you have a large dataframe, maybe access to
columns is exclusive? How would that interact with iterators - frame,
series, and series_vector alike?

Thinking about this a bit more, it almost seems like frame::corr()
should take a task worker as a parameter and the MP business can be an
internal concern for mainframe.