Moved metaprogramming machinery in frame.hpp into detail namespace
This is causing a strange problem with std::endl. I'm still
investigating this, but the compiler seems to want to
test/create/specialize a frame<> (no columns) in the course of
creating/specializing std::endl. I made a small change to add_all_opts
and remove_all_opts to allow for a frame<>, and also put all of these
classes into a namespace to avoid polluting the interface. I might
actually put these in their own header, like frame_detail.hpp or
something.

Of course, this isn't a "fix" - I'd still like to know why this was
interfering with std::endl. But msvc doesn't have very useful error
messages when compiling this. Nevertheless, I think it's still good to
be able to have a frame<> object, if for no other reason than there's no
conceptual reason why it shouldn't work and the underlying tuple can be
a tuple<>.

All of this has been on the road to figuring out why std::cout seems to
be broken (hangs) when mainframe_test.exe is run in the debugger.

TODO:
- Figure out std::cout and the visual c++ debugger

- CSV parser. Not 100% sure what the interface for this should be - but
  I'm thinking now that it should actually return a frame<Ts...> or
  perhaps a frame<mi<string>>. uframe should really stay a proxy object.
  In fact I'm considering putting uframe into the mf::detail namespace.

- SIMD-optimized correlations. One complication here is that SIMD
  support with msvc is slightly different than on g++, wrt compile-time
  vs run-time support.

- Join functions - frame::innerjoin, frame::outerjoin, frame::leftjoin.

- frame::groupby(). I'm still not 100% sure how to do this, but I think
  the key will be to create a new class type, like

  template< size_t ... indexes, typename ... Ts > grouped_frame;

  This would be returned by frame::groupby() and would include the same
  columns tuple that frame does (our usual non-mutating return-a-frame
  operation), but also maybe a multimap or unordered_multimap that
  indexes into the rows.

  I'm not sure how much frame functionality grouped_frame should take
  with it - also, should index values be exact or should grouping allow
  for ... groups?

- Add a function call wrapper for function calls in expressions

  auto f2 = f1.new_column<double>( "ceil", _f( std::ceil, _0 ) );

- Multiprocessing locks and guards. I still need to think about this a
  bit harder. Would it be enough to just say that frame and series
  objects should never be shared by reference and instead should be
  "copied" between threads?