CSV parsing, and frame::row_type
frame is a bit analogous to tuple (and is in fact implemented as
containing a tuple) in that each column is a unique type and iterating
through the columns isn't really supported because the columns have
different types. I think that in general this is the right choice if
we're comparing it to the idea of type erasure, but it does make dealing
with rows complicated and clumsy. In the meantime, we have a frame_row,
but it's kind of an iterator illusion - it's actually just a container
of iterators pointing to the series_vector's inside the frame.

When actually dealing with a parser like the csv loader, it's painful to
deal with an indeterminate number of rows. We're already limited
somewhat by the type of frame - we need to know ahead of time when
parsing the csv how many columns it has, and then return that many
columns regardless of what was actually in the csv. That can sort of be
mitigated by returning an oversized frame<mi<string>...> and then just
culling columns that are all missing, but in the parse itself, it's
annoying to have to deal with the indeterminate number of columns. For
the time being, frame_row is now frame_iterator_row and I'm introducing
frame::row_type, which is actually a vector of variants of all the
column types in the frame. I don't quite know whether this is the right
approach, though.

Right now, it has a series problem - if we have

using frame_type = frame<double, mi<bool>>;
frame_type f1;
frame_type::row_type row;
row.push_back( 1.0 );
row.push_back( true );
f1.push_back( row );

...believe it or not, row.push_back( true ) actually ends up adding a
double to the row, not an mi<bool>. C++ doesn't like
transitive-type inferencing - in the case of row.push_back( true ), it
needs to infer that this is bool -> mi<bool> -> variant<mi<bool>> and
instead prefers bool -(cast)-> double->variant<double>.

Can this be fixed? In the case of all mi<string>'s like the csv parser,
this is completely harmless because all the column types are the same.
In the case of heterogenous frame columns, it means in the documentation
I'll have to make it very clear that you need to specify row.push_back(
mi(true) ) rather than just row.push_back( true ).

All of this is a bit unergonomic. I'm still not quite sure how to handle
this. I'm now reconsidering maybe making uframe a bit more capable and
exposing that as a front-line type.

TODO:
- Finish CSV parser, handling special cases. parse_line() state machine
  needs to be refined a bit for escape sequences (or what passes for
  escape sequences in CSV. Also, utf8. utf8 is completely unhandled.
  Completely. There needs to be a utf8 scanner inserted into
  parse_line().

- SIMD-optimized correlations. One complication here is that SIMD
  support with msvc is slightly different than on g++, wrt compile-time
  vs run-time support.

- Join functions - frame::innerjoin, frame::outerjoin, frame::leftjoin.

- frame::groupby(). I'm still not 100% sure how to do this, but I think
  the key will be to create a new class type, like

  template< size_t ... indexes, typename ... Ts > grouped_frame;

  This would be returned by frame::groupby() and would include the same
  columns tuple that frame does (our usual non-mutating return-a-frame
  operation), but also maybe a multimap or unordered_multimap that
  indexes into the rows.

  I'm not sure how much frame functionality grouped_frame should take
  with it - also, should index values be exact or should grouping allow
  for ... groups?

- Add a function call wrapper for function calls in expressions

  auto f2 = f1.new_column<double>( "ceil", _f( std::ceil, _0 ) );

- Multiprocessing locks and guards. I still need to think about this a
  bit harder. Would it be enough to just say that frame and series
  objects should never be shared by reference and instead should be
  "copied" between threads?