reverse_frame_iterator and const_reverse_frame_iterator. Mostly.
I changed frame_row a bit so that it's a typedef of _base_frame_row
which has a dummy template type argument. This is so it looks a bit more
like _row_proxy - if the two types share the same type argument
signature, then can be interchanged with template-template type
arguments and I don't have to duplicate logic for operators.

std::sort crashes when sorting a mf::frame so we're not finished just
yet. But we're close.