Getting rid of csv parser.
In some ways, having a csv parser included in the library makes a lot of
sense because it increases the value of the library for users. But csv
parsing turned out to be a bit clumsy. mainframe's csv parser would have
looked a bit like this:

0. create an empty frame<Ts...>
1. parse a csv file line into a bunch of strings
2. convert/parse the individual strings into other types Ts... (int,
   double user types, etc)
3. insert a new row into the frame<Ts...>
4. repeat for all lines in csv file

The conversion in #2 needs to be very flexible because mainframe itself
is completely flexible when it comes to handling column types. This
necessary flexibility placed in the middle of the parser - after line
parsing but before insertion into the frame object - makes a general csv
parser that outputs a complete frame object a bit awkward to say the
least.

The real value of a csv parser isn't doing the very application-specific
and customizable element parsing (#2) - rather, the value is in the
state machine that's necessary to parse a line (#1) in a csv file
(considering all of the silly ways that csv escapes special characters).
That - parsing a line - is VERY useful, but also isn't really specific
to mainframe. So I'm removing it. Maybe I'll make a nice little
stand-alone csv parsing library.

TODO:
- How are frame_iterators coercing to each other? This shouldn't happen
  but perhaps something with uframe is causing this. This is really
  wrong.

- SIMD-optimized correlations. One complication here is that SIMD
  support with msvc is slightly different than on g++, wrt compile-time
  vs run-time support.

- Join functions - frame::innerjoin, frame::outerjoin, frame::leftjoin.

- frame::groupby(). I'm still not 100% sure how to do this, but I think
  the key will be to create a new class type, like

  template< size_t ... indexes, typename ... Ts > grouped_frame;

  This would be returned by frame::groupby() and would include the same
  columns tuple that frame does (our usual non-mutating return-a-frame
  operation), but also maybe a multimap or unordered_multimap that
  indexes into the rows.

  I'm not sure how much frame functionality grouped_frame should take
  with it - also, should index values be exact or should grouping allow
  for ... groups?

- Add a function call wrapper for function calls in expressions

  auto f2 = f1.new_column<double>( "ceil", _f( std::ceil, _0 ) );

- Multiprocessing locks and guards. I still need to think about this a
  bit harder. Would it be enough to just say that frame and series
  objects should never be shared by reference and instead should be
  "copied" between threads?