Adding the grouped_frame module. I'm not really sure how to handle this.
How exactly should I implement groupby() functionality? It seems like I
should return a new frame object ... or not? If not, I could implement
this using expressions behind the scenes. But this isn't a true index
and wouldn't really perform very well.

I think there are two problems here. First is that we want to be able to
group, and by that count we want to have an index. Then, we want to be
able to perform a function like mean or count or max or something on the
groups.

The second problem I think is straight-forward - we use whatever index
we build to perform a function and then return a new mf::frame with the
results. It's a bit complicated - for the data frame

  |       date | temperature |  rain
__|____________|_____________|_______
 0| 2022-01-01 |         8.9 | false
 1| 2022-01-02 |          10 | false
 2| 2022-01-03 |        11.1 |  true
 3| 2022-01-04 |        12.2 | false
 4| 2022-01-05 |        13.3 | false
 5| 2022-01-06 |        14.4 |  true
 6| 2022-01-07 |        15.5 | false
 7| 2022-01-08 |         9.1 |  true

if we do something like

    auto f2 = f1.groupby( _2 ).mean( _1 );

this should return a data frame like

  |  rain | mean(temperature) |
__|_______|___________________|
 0| false |             9.983 |
 1|  true |            11.533 |

But do we want to support something like

    auto f2 = f1.groupby( _2, _0 ).mean( _1 );

The second problem is what gets returned by

    f1.groupby( _2, _0 )

It's tricky to declare a template with multiple parameter packs. For
example, you can't declare

    template< size_t ... indexes, typename ... Ts > grouped_frame;

...which means that I'll have to be more imaginative about this.

TODO:
- frame::groupby()

- Join functions - frame::innerjoin, frame::outerjoin, frame::leftjoin.

- Add a function call wrapper for function calls in expressions

  auto f2 = f1.new_column<double>( "ceil", _f( std::ceil, _0 ) );

- Multiprocessing locks and guards. I still need to think about this a
  bit harder. Would it be enough to just say that frame and series
  objects should never be shared by reference and instead should be
  "copied" between threads?