Changes to get groupby() working
I've largely decided that the original frame<> is contained in
group_frame<>, and then we rearrange the columns first for indexing
(where we remove all other columns) and then later for returning
results. This is cheap and easy given that columns are
reference-counted.

Still, I really feel like there's something here that needs to be fixed.
In particular, it's not clear how exactly the functions should be
implemented and it makes me think that maybe there's something missing
in the interface? If I'm running

  group_frame<index_defn<0,1>,frame<month,float,bool,long>>::count()

...how is this implemented? You have to construct a frame of

  frame<month, float, int> out; // int is for count

Then we go through the index which gives us lots of (month,float)->row
num mappings for a particular (month,float). Rather than use the row
num, we just increment a count. Once we've moved on to the next
(month,float), we tally the count and are supposed to insert
(month,float,int) into out. But that's kind of clumsy when we generalize
on the index (month,float) - we have the index_key data structure to
represent this, but we need to push_back with out not only the index but
the count ((month,float),int).

It gets even clumsier with something like max( columnindex<Ind>... )
because then we need to find the max for multiple columns. We do this by
going through the index to get (month,float)->row num for a particular
(month,float), and then we look up that row in the original
frame<month,float,bool,long>, and then we need to perform an operation
on... what? Well if it was invoked as gf.max(_3), then just the last
column of long's. But it could be gf.max(_2, _3) as well in which case
we have either
  frame<month, float, long> out; // gf.max(_3)
or
  frame<month, float, bool, long> out; // gf.max(_2, _3)

...and then how to iterate over these? And how to do the insertion to
out?

It feels like there's some reduction or fold function that we should
have on frame that would make this easier. It's almost as if joins would
help here.

Anyways, what I have so far is

- index_defn<> and associated transformers (*_contains<>, *_deref<>,
  *_select<>, *_prepend<>, *_invert<>)
- indexed_frame<>
- result_frame<>
- got rid of pointerize<>
- README changes