grouped_frame<>::sum()
This was a bit tricky because the build_index() method had a defect that
caused it to be accumulated each time it was called. This made no
difference to min() and max() but sum() ended up summing up rows
multiple times.

Next will be mean() and then stddev(). mean() will probably require a
new rowop function and stddev() almost certainly will.

Going forward, sometime in the future, I guess the question is should
this indexing functionality just be folded into frame<> in some way? It
would have to be type-erased of course but maybe that doesn't matter so
much for an index. This would allow neat pandas-like functionality like
re-indexing the frame with a new series. But that's quite advanced and I
don't think I actually need that functionality myself.