I think I've put operator[] in it's right place
Previously I wasn't really sure how operator[] should work with
mainframe. I had defined frame::operator[](size_t) which returned a
_row_proxy, but then that left a couple of questions.

1. Can we define frame<Ts...>::operator[](Ex) where Ex is a row selector
   as a front-end to frame::rows(Ex)? This would allow syntax
   like
     auto f2 = f1[_2 > 2022y/1/1];
   And...
2. Can we define frame::operator[columnindex<Inds>...] that works as a
   front-end to frame::columns(columnindex<Inds>...)? That would allow
   syntax like
     auto f2 = f1[_3, _2, _1];

It turns out that #1 is pretty easy, kind of. But #2 is really hard
because up until c++23 the language doesn't support multiple arguments
to operator[]. Traditionally to facilitate that developers have created
proxy objects that implement the comma operator (operator,). So that's
what I had to do. The new proxy type is columnindexpack<Inds...> and
this is what allows the #2 syntax.

Complications? Yes, there are a few.

a. The syntax
     auto f2 = f1.rows(_2);
   ...is completely ok as long as column _2 is bool convertible. That
   kind of means that
     auto f2 = f1[_2];
   ...should be ok too, but of course this is also a column selector. So
   we won't allow this and instead will require
     auto f2 = f1[_2 == true];
   ...for this case.
b. Both frame::operator[](columnindex<Inds>...) and
   frame::operator[](Ex) both return a frame object. But
   frame::operator[](size_t) returns a _rowproxy<> object. That is very
   much not OK so I've banished frame::operator[](size_t) and instead
   this will have to be accomplished with frame::row(size_t)

Anyways, frame::operator[]() is now the all-purpose
slice-and-dice-this-dataframe tool for cuttine and rearranging rows and
columns.