frame::prepend_column() wasn't as easy as I thought it was
My initial copy-and-paste implementation of frame::prepend_column() was
very very simple but also very wrong. For one thing, it was actually
returning a frame<Ts..., T> rather than the correct frame<T, Ts...>. For
another, the expression that frame::prepend_column() takes refers to
columns but the column numbers change after prepending the new column.
Should the expression columns refer to the old or the new numbering?
They should almost certainly refer to the old numbering because that's
just much more intuitive - the function hasn't returned yet and it seems
strange that it would be operating on a column numbering that the user
hasn't even received a frame for. So I needed to make modifications so
that the old numbering would be valid. So for example, with

    frame<year_month, double, int> fr1;
    fr1.set_column_names( "date", "coef", "indx" );
    fr1.prepend_column<short>( "newval", _2 + 10 );

... without this fix the expression would populate "newval" with "coef"
+ 10 but with this fix the expression will populate "newval" with "indx"
  + 10.