Added expression functions
I'm not sure that I'm 100% happy with this design, but I'm not sure that
I could really do any better. Previously, if you wanted to call
frame::prepend_column(), frame::append_column(), frame::rows(), or
frame::operator[] with an expression, you'd be limited to statements
like

    fr[ _1 > 20 ];

If you wanted to actually test the rounded value of _1, OTOH, then you'd
have to use a lambda:

    fr.rows([]( const auto& b, const auto& c, const auto& e) {
        return ::round( c.at( _1 ) ) > 20;
    });

Nothing wrong with that, but it's no longer a 1-liner. With this new
functionality, you can now call

    fr[ fn( ::round, _1 ) > 20 ];

For now it's super-fragile. For one thing, it really can't deal with
overloaded functions. And the error messages that it returns in that
case are really messy. I'd like to figure out a static assert for that
case.

Also, I haven't added a static assert for the case where the user hasn't
provided enough or has provided too many parameters for the function.

But it does work.