Added frame::operator*
Surprisingly, I haven't had to use this up until now. Or rather I've
used iterators which are an easier way of getting at the frame data.

This allows you to use syntax like

    frame_row<year_month, int, double> fr1;
    fr1.push_back(2022_y/11, 1, 3.14);
    fr1.push_back(2022_y/12, 2, 2.718);
    year_month a = *fr1[_0];
    year_month b = *fr1[1][_0];
    year_month c = *fr1[_0][1]; // same as b
    int d = *fr1[_1];
    int e = *fr1[_1][1];
    int f = *fr1[1][_1]; // same as e

Note that b, c, e, and f can be very inefficient. With mainframe, column
selection is fast and easy. But row selection involves cutting. This is
the cost of supporting sophisticated row selection syntax - the general
case isn't a single element but rather a new frame with the selected
rows.